Support H2 for some evaluations
The database schema should now migrate successfully for both h2 and
postgresql. The common-to-most-evaluations queries succeed on both
h2 and postgresql, allowing some evaluations to complete successfully.

A summary of changes included (not exhaustive):

* project_id and source_id promoted to long (will increase heap use)
* Cache/ORM ids promoted to long.
* Project table stores content id in hash column.
* New changesets for several tables.
* ExecutionLog tracks end_time, ProjectExecutions shows duration
* App inserts project_name into ExecutionLog.
* New class ExecutionResult to support the project_name.
* Updated JUnit system tests to account for ExecutionResult.
* Comment in wresconfig.xml showing how to use H2.
* Moved app-level partition methods into the Database class.
* Moved log execution method into the Database class.
* Added bulk insert when trying to copy to non-postgresql database.
* Refactored copy to account for the above change, simplified it.
* Made DatabaseLockManager accept long instead of int.
* Added DatabaseSettings jdbcUrl and host, deprecated url.
* Removed unused code.

Refs #60801

+++ wres-io/src/wres/io/data/caching/Features.java
@@ -69,17 +69,17 @@ public class Features extends Cache<FeatureDetails, FeatureKey>
 	 * @return The ID for the specified feature
 	 * @throws SQLException Thrown if the ID could not be retrieved from the Database
 	 */
-    public Integer getFeatureID( FeatureDetails detail )
+    public Long getFeatureID( FeatureDetails detail )
             throws SQLException
 	{
 	    LOGGER.trace("getFeatureID - args {}", detail);
 		return this.getID(detail);
 	}
 
-	public Integer getFeatureID( FeatureKey key ) throws SQLException
+	public Long getFeatureID( FeatureKey key ) throws SQLException
     {
         LOGGER.trace( "getFeatureID with FeatureKey arg {}", key );
-        Integer result;
+        Long result;
 
         synchronized ( this.keyLock )
         {
@@ -89,7 +89,7 @@ public class Features extends Cache<FeatureDetails, FeatureKey>
                               key );
                 FeatureDetails featureDetails = new FeatureDetails( key );
                 this.addElement( featureDetails );
-                Integer fakeResultForLRUPurposes = this.getID( key );
+                Long fakeResultForLRUPurposes = this.getID( key );
                 result = featureDetails.getId();
             }
             else
@@ -117,7 +117,7 @@ public class Features extends Cache<FeatureDetails, FeatureKey>
             throws SQLException
     {
         LOGGER.trace( "getFeatureKey called with {}", featureId );
-        for ( Map.Entry<FeatureKey,Integer> cacheEntry :
+        for ( Map.Entry<FeatureKey,Long> cacheEntry :
                 this
                         .getKeyIndex()
                         .entrySet() )
@@ -401,7 +401,7 @@ public class Features extends Cache<FeatureDetails, FeatureKey>
     }
 
     @Override
-    Integer getID( FeatureKey key )
+    Long getID( FeatureKey key )
     {
         LOGGER.trace( "getID( FeatureKey={} )", key );
         synchronized ( this.getKeyLock() )
@@ -409,10 +409,10 @@ public class Features extends Cache<FeatureDetails, FeatureKey>
             LOGGER.trace( "getID( FeatureKey={} ) key.hasPrimaryKey() was false, matching on partial key...",
                           key );
             // Otherwise, scan the keys for a match on a partial key
-            Integer id = null;
+            Long id = null;
             boolean foundIt = false;
 
-            for ( Map.Entry<FeatureKey, Integer> keyId : this.getKeyIndex().entrySet() )
+            for ( Map.Entry<FeatureKey, Long> keyId : this.getKeyIndex().entrySet() )
             {
                 if ( keyId.getKey().equals( key ) )
                 {
