+++ wres-io/src/wres/io/retrieval/ObservationRetriever.java
@@ -163,7 +163,7 @@ class ObservationRetriever extends TimeSeriesRetriever<Double>
             }
 
             // Existing units
-            int measurementUnitId = provider.getInt( "measurementunit_id" );
+            long measurementUnitId = provider.getLong( "measurementunit_id" );
 
             // Units mapper
             DoubleUnaryOperator mapper = this.getMeasurementUnitMapper()
